Move domain worker assignment to migration planning (from apply).
During a migration, we currently decide what workers to place
domains on in `migration::apply`. By moving this earlier in the
migration flow, to migration planning, we can catch migrations
that we cannot satisfy earlier and more easily prevent
putting the system in a inconsistent state due to failures
during migration application.

This CL moves domain assignment to the `plan` stage of
migrations. Each shard of each domain is assigned a worker,
exactly the same as before, and this mapping is passed
in the MigrationPlan to the `apply` stage.

Change-Id: Icc502a87616f42dd55441510b7b675251157ae3c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/153
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>