Use ECR in CI override, not by default.
Currently, signing into ECR in the build account is a bit of a mess and
requires a ton of setup. We need to use ECR for Buildkite since
otherwise we get throttled by Docker Hub. But as individual users, we
should not be producing enough traffic to become throttled.

Also, having all the image locations closer together makes it clearer
that all are being overridden. Consul was not being pulled from ECR.

Change-Id: Iec26649717c49a0516c21b3a0383aba846d54532
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/157
Reviewed-by: Griffin Smith <griffin@readyset.io>
Tested-by: Buildkite CI