Remove query sanitization from noria-client
From reading through git history, it looks like the hacky regex-based
query "sanitization" that noria-client was doing - specifically removing
comments, folding whitespace, and converting double to single quotes -
was a known hack that was added to pass a demo, and was never intended
to be a prime-time feature. Specifically, in this instance,
ENG-164 (MySQL adapter does not correctly insert strings containing
double quotation mark characters) was caused due to the string quote
replacement being generally over-zealous. Removing sanitization causes
that ignored test to pass, and doesn't break anything *else*, so should
be fine to do for the time being - if we want to bring back ignoring of
comments during parsing, we should do that in the *parser* instead, not
here.

Change-Id: Ibf63717965bbcee3a210f9765c3154973668de26
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/156
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>