Make dialect a runtime flag, not a feature flag
Convert the configuration of which dialect we use when parsing SQL from
a feature flag to a runtime flag, an enum passed around through
basically all parse functions in nom-sql. Rust feature flags were never
really intended to be used for this kind of thing (selecting conditional
behavior), and are instead intended to be used for *additive* features
on a crate where their addition doesn't actually break anything that
would be dependent on their absence. Concretely, cargo unifies all
feature flags for a particular crate dependency across all crates with
that dependency in the workspace, meaning as soon as we tried to
actually *depend* on nom-sql with the postgres feature flag enabled in
one crate (noria-psql, in I50d32e90) *every* crate in the workspace got
that feature flag enabled, causing things depending on the mysql dialect
to break. There are workarounds - for example, we could move noria-psql
out of the workspace - but despite the size of this change it was
relatively mechanical to make and felt like the most correct and
idiomatic way of achieving what we want to achieve.

As of this change, the actual selection of dialect is hardcoded to MySQL
at the very top level of the parser in `parse_query_bytes`, but
subsequent changes will make it an argument to that function and
`parse_query`, then update the various callsites to select which dialect
they want. The idea for noria-server, on the other hand, is to have a
single canonical dialect that all clients and adapters translate
to (MySQL seems fine, since that's what Noria was originally written to
use), so that's added as a constant in the recipe code and documented
there.

Change-Id: I564dd89e243703220b468b8781bb368dc839de82
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/167
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>