Add a Dialect argument to parse_query
Allow passing the Dialect to use when parsing to the top-level call to
parse_query in nom_sql, and thread this through the various places that
call it so that the correct dialect is used to parse in the correct
places. This includes a new `dialect` field on
noria_client_adapter::Adapter, noria_client::BackendBuilder, and
noria_client::Backend, which is set to MySQL for the mysql adapter and
PostgreSQL for the postgresql adapter. Since users might be using either
dialect in their applications, the dialect to use for parsing is
user-configurable in the js-client, using a new "dialect" key to the
backend options, which can be either "mysql" or "postgresql",
case-insensitively.

Change-Id: Ic3d0a3042733de162bb1e1d579acaedda41deeab
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/171
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>