Dependency-invert upstream database connection
Currently, the noria_client Backend is pretty tightly coupled to MySQL
specifically, via embedding a hardcoded MySqlConnector, and using a lot
of msql-srv-specific types all over the place - we'd like to break that
tight coupling, by making it completely agnostic to the specific
upstream database being used for passthrough and fallback, *and* the
specific client protocol being used for serializing schema and results.

This commit is the first of probably many steps in breaking that
coupling - it introduces a new "UpstreamDatabase" trait, which
essentially replicates the API used by MySqlConnector and
PostgreSqlConnector, and is implemented by both of those types. The
Backend itself now has an additional type parameter `DB` which is
constrained to this trait and is used by Reader and Writer as the type
for the upstream database connection. The QueryResult enum which is
returned by multiple methods on Backend now uses two associated types on
UpstreamDatabase (ReadResult and WriteResult) as the types of the
results returned when performing passthrough writes or fallback reads.

The eventual goal here is that we make the noria-client crate completely
unaware of mysql *or* postgresql, but that hasn't happened yet as of
this commit, since both MySqlConnector and NoriaSqlConnector are both
still defined inside of noria-client - in a follow-up commit I'm going
to be moving those two types to their respective binary crates, in
addition to the implementation of msql_srv::MySqlShim for Backend<A,
MySqlConnector>

Change-Id: Ibd5ba06cf7c790168d3f01be653df8127934ebd6
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/177
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Reviewed-by: Peter Farr <peter@readyset.io>