Make write_query_results a top-level function
Generally accepted rust style, as I understand it, is to limit
associated functions on a type to functions that actually *directly*
depend on that type, either by including a `self` receiver or having
Self mentioned somewhere in the type. The write_query_results function
on Backend doesn't fit this guideline, and hence required explicitly
specifying an instantiation of the `A` type on Backend to call it, even
though it didn't care at all about what that type was instantiated *to*.
This gets even more annoying when we want to add new types parameters to
Backend, since it requires updating all the call sites, again without
actually having any effect. Promoting it to a bare top-level function is
more idiomatic, and much easier to work with overall.

Change-Id: I4a274f7f9a91b2c1ff9645849d1f4b4b0b1d181b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/179
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>