Refactor and start work towards validate all
The first step of this is to make a task that runs validate over all the
different roots that Substrate generates. This can be highly parallel.
Once all of them pass, then we can run parallel plans for branches/CLs
while only running the applies on main.

The single file was getting to large and so I broke up the runner into
multiple files that focus on different parts.