Update ami_account_id for build account
Point the ami_account_id of the readyset terraform module at the build
account, which is where packer images are going to be sent for now. At
some point in the future this will likely be configurable (so that we
can have customer-facing release AMIs in a different account) but for
now we're only deploying internal clusters so continuing to hardcode it
should be fine.

Change-Id: I033f0dff462bf5fd883e674ad7b36514faf27dc2
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/187
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>