Add release automation for psql adapter
Add a buildkite step to build a release binary of the psql adapter, and
a packer image for the psql adapter that's a direct 1:1 clone of the
packer image for the mysql adapter with all occurrences of mysql
replaced with psql (except in comments, where it's replaced with
PostgreSQL). Because of our deploy automation for packer image builds,
just adding the directory to ops/images is enough to get the packer
image building.

Change-Id: I4d1e810bbca1e74b7d8474cb44209e8eea1b14d1
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/188
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>