Fix empty resultset on fallback
This fixes a bug where if we got no results from fallback, we would
simply write an ok packet. What we actually need to do is write a header
row with the columns that were requested in the original query, and then
an EOF packet.

I've captured and confirmed this with wireshark. This commit fixes this
functionality.

This was initially caught in the sequelize framework test. I've
confirmed that this fixes this functionality and gets us past this
error in the framework test.

In the future we should probably just pass the packets back that we get
from MySQL directly to the caller with zero interpretation or
modification.

Change-Id: Ifbdc37a79b8171a3541fb45e872f9cb34eda1fdb
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/192
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>