Add postgres support to noria-logictest
Add support to noria-logictest for conditionally using either postgres
or mysql as the adapter to run, the database to replicate from if
running with replication, or the database to run logictests against if
not running against noria.

To accomplish this, this commit breaks out a number of enums around
dynamic connections, URLs, and database types that can be either mysql
or postgres into a new `upstream` module in noria-logictest (which I'm
planning on putting into a separate crate, since it'll be useful for
things like demo code and also replicators), and then uses this instead
of anything mysql-specific in the runner code. ast::Value also gets
instances of postgres's FromSql and ToSql traits, and the command-line
arguments have been reworked both to be more URL-centric, and to not
specifically say "mysql" anywhere.

Change-Id: I7853d2dc2cbc4c46d0f748c26cda04c821af56ae
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/193
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>