Add a user offset parameter to fastly demo
Since we want to be able to pass different key subsets to different
readers, this flags allows to provide the index of the minimal key
to use. For example if we have 100000 keys, one can speicify 50000
for the user-rows, and 0 for the offset for one reader and 50000
user-rows with a 50000 offset for the second reader, so reader one
handles the [0,49999] range and reader two the [50000,99999] range.

Change-Id: Ibef7d43d513f965eb31a82339a014a0987cfbd11
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/197
Reviewed-by: Justin Miron <justin@readyset.io>
Tested-by: Buildkite CI