Remove pool from mysql_connector
The connection pool is removed from mysql_connector and replaced with a
single mysql_async connection to the mysql server. This is done because
the adapter <-> server connection contains state (e.g., a table of
prepared statements). A single client <-> adapter connection should
maintain visibility of this state throughout its lifetime. This is not
guaranteed with a connection pool.

race_reads() now creates its own connection on each call instead of
grabbing one from the pool used by the existing reader mysql_connection.

Change-Id: I120fabac6b73e97bce042acddcedd2cb0c8a4fdb
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/199
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>