Fix panic when evicting from an empty KeyedState
Previously, an attempt to call `evict_with_seed` on an empty HashMap
KeyedState would panic due to the use of modulo with the right-hand side
value equal to 0. There was already a check for all the BTreeMap
variants of KeyedState (unused, currently) to make sure they weren't
empty before we did that, so this commit just duplicates that check to
all the other variants. There's also a (originally failing) test
covering this for *all* variants of KeyedState, including all lengths
and both IndexTypes, just for good measure.

Change-Id: I2624f8d0768a63f85da941e5c2f462db565e33e6
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/203
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>