Packer + Ansible Image Generation for Tailscale
Picking the most obvious tools for doing this work comes to Packer and
Ansible. We were previously using inline scripts but as these images
become more complicated or we want to abstract over different
implementation details, not having to write it all ourselves will be
useful.

Currently, the Tailscale nodes are using cloud-init and pulling down a
tarball of Ansible source code to provision themselves on boot. This
does not seem like a great strategy going forward and this gives us a
reasonable place to start experimenting with creating images for our
software in general.

To actually be able to build these new images, this required creating an
ops image that is used to build the various tools needed for both
linting and doing the operations. As such, this cleaned up a number of
previous steps.

Change-Id: I40956fe0e12c28fa91ebc377349b6e3fb3e840c3
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/204
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>