Refactor `reader_direct` to enable multiple database types.
This change abstracts query generation to a separate
`QueryExecutor` struct. This not only makes this code much
cleaner, it allows us to use the same client to support different
target databases. A future change will incorporate a mysql client
into this abstraction and as a result make having reader.rs and
reader_direct.rs unnecessary.

Change-Id: I8f8e06a82d188939fdb74d94788adb53e34f3629
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/205
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>