Allow selecting adapter for noria-client tests
Allow selecting which adapter to use in the test utils created for the
noria-client integration tests. Since the types we have to return for
connection opts are different, this is implemented as a trait with some
associated types (the instance of which we have to explicitly pass to
common::setup() with a turbofish, since it's not actually instantiated
so can't be inferred).

Change-Id: Id7e0e50eaca3523735c2f93eb35dd5d922ee4cf3
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/211
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>