Add (ignored) tests for postgres value round-trip
Add a new integration test to noria-psql, with a set of macro-generated
proptests for round-tripping values of different types to the database
in various different ways, including:

* as an inserted value
* as a parameter
* returned from a query
* through fallback

This can be extended to even more postgres types in the future, and
should start to give us confidence that our implementation of postgres
type conversion both via psql-srv and through fallback is correct.

As of this commit, all of these tests are failing, so all are marked as
ignored - but this should give us a starting point to start getting all
of these types working.

Change-Id: I5d6d0552dcfb1164aded297519138145474768d7
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/215
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>