Added `Ord` and `PartialEq` cases for complex types against `Text` types.
We were comparing `left` and `right` on `Op` expressions, without checking the
types.
This caused an issue with `DateTime` types (and I'd say, the issue will also
affect any other type that has to be parsed from a `String`), because we were
comparing `Timestamp` against `Text`, which always yields `false`.
Now it is possible to compare (by `Ord` and `Eq`) `Text` and `TinyText` against
`Timestamp` and `Time`.

Refs: ENG-515
Change-Id: Ic15f0c70adc32306bb483ab54aa3008cd7efc887
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/220
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>