Make SqlType lengths optional
In MySQL, the length fields that are supplied to multiple types are
usually optional (the exception, empirically, seems to be VARCHAR and
VARBINARY) - previously, we would parse these as optional, but then fill
in the default value for the particular numeric type in question *in the
parser* - preventing anything downstream *or* anything that was
constructing and passing around SqlTypes itself from being able to
encode an unspecified length as part of a type definition. This is
generally problematic from a style / separation-of-concerns
perspective (parsers generally should represent the semantic structure
of the code as written as closely as possible, without having any
business logic like default values), but actually becomes a real problem
once we start wanting to deal with postgresql, which doesn't support
lengths on integers *at all* - we need a way to represent column types
without lengths, if only so that things like query_generator can emit
CREATE TABLE statements that postgresql knows how to parse.

This commit turns all of the length fields inside of SqlTypes where
mysql treats the length as optional into Options, removing the setting
of default lengths entirely (it's never checked on insert, so it's fine
to just leave it as None if it wasn't specified). Also, both CHAR and
BINARY have optional lengths in mysql (tested empirically) but we
weren't even *parsing* those without lengths, so this updates those to
be optional as well.

Refs: ENG-488
Change-Id: Ib3be7d0688d490f06e976c603759348f4538fd2c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/221
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>