Feature gate RYW
Currently we get back the wrong result from MySQL if the
CLIENT_SESSION_TRACK capability flag is enabled. We use this flag for
RYW to attach extra information onto the OK packet that we get back from
MySQL.

This commit gates RYW behind a compile time feature of the same name, so
that we can enforce correctness until we decide that RYW is production
ready.

Change-Id: Ic69157667fd1ef9a87b9c635b6d856c8e77bfaa6
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/223
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>