Move connectors to the adapter crates
Another step in the ongoing quest to decouple noria-client from mysql:
here, we move mysql_connector to noria-mysql and postgresql_connector to
noria-psql, replacing the word "connector" with the less generic
"upstream" in the process, for consistency with the UpstreamDatabase
trait that they both implement. Similar to what noria-psql was already
doing, noria-mysql now has a new Backend newtype that wraps the
noria_client Backend fixing the type parameter to MySqlUpstream, and
uses this to actually implement msql_srv::MysqlShim.

The integration test suite in noria_client, which was previously
mysql-specific, has also moved to noria-mysql, and the common setup
code for noria-client integration tests which was added back in
a8635b1ee (Allow selecting adapter for noria-client tests, 2021-08-24)
has moved to a new top-level `test_helpers` module in noria-client. I
tried to put this behind a feature flag so we don't compile it in
release builds, but was unable to get the dev dependency stuff working
for noria-mysql and noria-psql - this probably doesn't actually matter,
as LLVM is pretty good at dead code elimination anyway.

Very little else of substance is happening here, despite the relatively
large diffstat - the bulk of this commit really is just shifting things
around and updating references to stuff.

Change-Id: I0aa0fd3834d46a909a44009b4929bbe9de713a51
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/227
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>