noria_client: Remove HARD_CODED_REPLIES
Remove the `static_responses` and HARD_CODED_REPLIES functionality from
noria_client and all related code. These were present as a workaround
for not supporting certain statements that an old version of rails
needed to run on connection, but now that we have proper support for
query fallback they're completely unnecessary.

Change-Id: I4367ab0ace453544332fece2538b0aafe80a4a1c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/228
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Reviewed-by: Peter Farr <peter@readyset.io>
Tested-by: Buildkite CI