Test basic adaptor deployment in main pipeline
Commits lately have been breaking adaptor deployment. This is our fault
for not having basic testing of adaptor deployment in the main pipeline.

This adds basic testing of adaptor deployment to our main pipeline as a
requirement for merging CLs. It uses a modified version of the same
docker compose setup we use for framework testing, but with a dummy
image at the end that runs a healthcheck on the adaptor.

Like our framework testing, the health checks are chained
(readyset-server requires mysql to be healthy, readyset-mysql requires
readyset-server to be healthy, and lastly our dummy image requires the
adaptor be healthy).

Change-Id: Ib8ca0e44db70cceb68fef96591bc1c8919c70ab5
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/231
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Michael Cronce <mc@readyset.io>