Fix rename of mysql-url -> upstream-db-url
As part of making noria-client upstream-db-agnostic (in preparation for
postgres support), the --mysql-url command line flag was changed to
--upstream-db-url - but not updated in a few scripts, docker images,
and (most importantly) another argument that requires it. This last one
is especially bad because despite not failing at compile-time, requiring
an argument that doesn't exist causes running the application to
instantly panic (boo, clap! I opened an issue[0] to request this to be
made a compile-time error). For that last case, I've also added a quick
unit test (failing, before the fix) to cover our command-line option
parsing so that stuff like this is less likely to happen in the future.

[0]: https://github.com/clap-rs/clap/issues/2740

Change-Id: Ie826c09400297545f3767163c32266e20ae731e9
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/232
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>