Only use jemalloc in the noria-server binary
Move the configuration of jemalloc as the global allocator from
noria-dataflow's lib.rs to noria-server's main.rs, so only the
noria-server binary uses it by default. This allows things that
transiently depend on noria-server to *not* have jemalloc as their
allocator, which is especially important for things like js-client that
need to dlopen our code (and dlopening jemalloc is, as
https://github.com/jemalloc/jemalloc/issues/937 puts it, "bad news
bears").

Change-Id: I05f260320b55a08d1c78bfa490fb109d8689ea14
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/234
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>