Installer for docker-credential-ecr-login
We are starting to run up against Docker Hub limits. As such, it makes
sense to move all of our Docker running against a registry we control.

This allows someone to run
`cargo xtask install-docker-credential-ecr-login`
and get their local Docker command to authenticate correctly against
the ECR registries and with no concerns about expiring Docker
credentials and instead relying upon AWS credentials.

This could have been some documentation but writing it as a script was
extremely easy and helps onboarding in the future.

Change-Id: I0623ca22ba6962a966afba9356212222c3eb8e29
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/238
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Justin Miron <justin@readyset.io>