Extract MysqlTime to its own crate
We'd like to invert the dependency that crates deep inside the noria
dependency tree have on mysql-specific functionality as much as
possible; a big part of this is removing as many dependencies on
msql-srv as possible (ideally, only noria-mysql itself would depend on
msql-srv). A lot of stuff throughout the dependency tree has
dependencies on msql-srv now, and it appears that the primary reason for
a lot of this is the need to talk about the MysqlTime struct, which
provides mysql-like semantics for representing Time values, and is used
as the Time variant in our DataType representation (for better or
worse). This struct stands pretty well alone from msql-srv, though, so
it makes sense to just extract it to its own top-level crate, and have
the stuff that previously depended on msql-srv depend on that instead.

A couple side notes:

* noria-dataflow depended on msql-srv, but seemingly only to re-export
  MysqlTime, and then nothing actually *imported* that re-export, so this
  was just removed wholesale
* noria and nom-sql still unfortunately both depend on msql-srv to do
  some schema type conversion stuff, this will be removed in a follow-up

Change-Id: Ic8b975dc80a4d17935642549ee0aae21319a21b3
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/240
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Tested-by: Buildkite CI