Creating mirror repositories in build ECR.
As a first thing to start managing with Terraform in the build account
without interfering with the current work, I decided to try and start
mirroring all of the images we use on Docker Hub to our own ECR. This
should also save a bit of money by reducing bandwidth costs.

This should set up all of the mirrors in a way that it is easy to add
another one if we need it. I have grepped through the main repo to find
every container currently in use. A CL is in progress to make a
Buildkite pipeline to keep these up to date.