Remove dep on msql-srv from nom-sql and noria
As the second part of the work that b21d7345d (Extract MysqlTime to its
own crate, 2021-08-27) was doing to remove the mysql-specific coupling
deep down in the crate dependency graph in the workspace, this commit
removes the the dependency that nom-sql and noria both have on msql-srv.
Mostly what this entails is shuffling schema conversion code around a
little bit - the `convert_column` function, which was previously an
associated function on nom_sql::ColumnSpecification, is now moved to
`noria_client::schema` (it will move all the way up to noria-mysql next
as part of the overall decoupling effort; there's a TODO to that effect
in the code) and the various column schema methods on
`noria::view::ViewSchema` now return nom-sql types instead of msql-srv
types, with the conversion to msql-srv happening in
noria_connector (this also will go away soon).

Change-Id: I8656ede78a09b5c3882aa0ac83bd6af84808bebe
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/241
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>