Decouple noria_client::Error from mysql+postgresql
The only remaining reason that noria-client had to depend on anything
specific to mysql or postgresql was the error type, which had to contain
variants for both mysql_async::Error and tokio_postgres::Error so that
the adapter binaries could represent errors with upstream databases. In
order to finalize the quest of making noria-client completely
upstream-agnostic, this makes the error type that we use an associated
type on UpstreamDatabase, and defines separate error enums with variants
for the relevant upstrams in noria-mysql and noria-psql, in addition to
variants for ReadySetError. Since noria_client now doesn't have an error
type of its own, all of the methods of NoriaConnector now just return a
ReadySetResult.

I played around a little bit with the design of this - initially I was
going to still have an Error enum in noria-client, and then have a
variant of that enum for UpstreamDB::Error similar to what we do for
QueryResult and PrepareResult. I ended up scrapping that approach
because due to rust's trait coherence rules that prevents us from
deriving a From impl from UpstreamDB::Error for that type, and the extra
error enum wrapping became prohibitively noisy - this approach ended up
being a lot cleaner, even if we do have to define two new error enums.

Change-Id: Ieb5970e51f184ade5219cd01f2ccc5f2aa0b0fd2
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/243
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>