Make env vars consistent
The dockerfiles we had for both readyset-mysql and readyset-server were
previously setting command-line flags via env vars passed to the docker
image itself, even though the binaries we build already support
configuration via env vars - and frequently, the env vars passed to the
images were *different* than the ones already supported by the binaries!
We're going to be deploying these docker images to production soon, so
it's better to just be consistent everywhere about the env vars we
support (for better or worse) - to that end, this commit removes the
setting of flags for the two binaries via environment variables, instead
opting to just let the binaries handle the env vars themselves, adding a
couple new env vars to binaries that previously only worked as
command-line flags, and updates all the places (that I could find, at
least) that were passing these env vars to use the same name.

Change-Id: I229474a3e0c4ac9cfc885fae3db216350aaccd10
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/246
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>