Remove version from all local dependencies
Having version constraints in workspace-local dependencies is kind of
pointless, since we're always just going to be depending on the version
that we have in the workspace. This hasn't caused us any problems yet,
since cargo seemingly just ignores the version constraints, but cargo
chef, which we're planning on using for our production docker images, is
choking on the version constraints. To get cargo chef working, this
commit removes all version constraints from all Cargo.toml files on
crates inside the workspace.

Change-Id: Ic823d42f6b40e151c2cf5723c1c131ff80118c12
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/249
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>