Implement QueryResult for Upstream Connector
This change puts the upstream database Read/WriteResult types inside a
QueryResult enum. This type is returned by the upstream database query
functions.

Prior to this change, the adapter had to know whether it was issuing a
read or write query when calling one of the upstream query functions.
This is problematic because the adapter is not able to parse all valid
queries that could be sent to the upstream database. In these cases, the
read or write result would not necessarily be returned to the client.

This CL will be followed by another to infer whether a read or write
query was issued based on the resultset from the upstream connector.
These two CLs will allow the adapter to issue upstream queries without
parsing the query being issued.

Change-Id: Iac9ae655f0991d2398dce82d171f3d8d745a40b9
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/248
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>