Generalize docker image building into a script
As we're making more and more changes to the way docker images are built
in CI, the number of places we have to make those changes has started to
become a little prohibitive - especially if we start to think about
optimizing caching for multi-layer builds etc. This commit generalizes
the process of building and pushing a docker image to our ECR repository
into a new build_image.sh script, and updates all places in CI config
where docker images are built to use that script.

Change-Id: Ib5ce98a0b60b83b667031212a3368e5865b9200d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/251
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>