Only tag latest image on main branch builds
It's possible that a build of a CL (not the main branch) breaks an image
somehow - for things that depend on `latest` we'd want to avoid pushing
such a broken image to the latest tag. This changes the build_image.sh
script (and hence all image builds in CI) to only tag images as `latest`
if the build is on the main branch.

Change-Id: Icf4ec6ad582b8cbc0142f7de58d2dfde48558edc
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/252
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>