Send transactions to fallback
We should be sending all transactions to fallback. This commit takes
care of that by handling the logic of starting a transaction, and then
either committing or rolling back that transaction when that is
requested of us by a client.

In the future we likely don't need this at all because in the case of
fallback we can just forward the raw packets back and forth.

This commit also includes a logic test to verify that transactions work
as intended. That is, that once a transaction begins we send ALL queries
to fallback, until we receive either a COMMIT or ROLLBACK signal, at
which point we exit transaction mode and proceed normal processing.

Change-Id: I8a4a8883a6da249865677dc2be6c1d82bbef1274
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/262
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Reviewed-by: Michael Cronce <mc@readyset.io>