noria-logictest from-query-log:  Add support for prepared statements
In 1b7f5801a661a2bf68c3a4ebe4d531070803b16b, `noria-logictest
from-query-log` was implemented with one fairly glaring omission:  It
ignored prepared statements in the querylog.  This is because, when
MySQL logs prepared statements, it logs the PREPARE call with the
parameterized query (e.g. `INSERT INTO t1 (c1, c2) VALUES (?, ?)`) as
expected.  However, when it logs the EXECUTE call, it doesn't just log a
reference to the prepared statement and the variables; it logs the
complete parameterized query, e.g. `INSERT INTO t1 (c1, c2) VALUES (1,
2)`.

To get prepared statements working, we had to _basically_ implement a
rudimentary query cache.  We parse the prepared statement, and store the
AST in a HashSet; on each EXECUTE call, we parse the query, replace all
the literals in the query with ? placeholders, then search the query
cache for a matching prepared statement.  If we find one, we execute the
prepared statement with the literals we pulled out in the previous step.

I am very thankful for what we already have in the form of nom_sql and
other `noria-logictest` bits; this would have been far worse without
them.

Change-Id: I9b8f3cc318bff411057d2d1798f20f12d3bdcf39
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/263
Reviewed-by: Griffin Smith <griffin@readyset.io>
Tested-by: Buildkite CI