New Image Builder, removes Ansible
After experimenting with Ansible, I threw it away and redid it all with
just shell scripting. As part of that, included shellcheck as a linter
for the variety of shell scripts we are going to write here.

Ansible added a huge amount of unneeded complexity for the level of
complication we want to be at. This also reduces the number of
dependencies we might implicitly end up depending on.

Left the current images in place for now to get this reviewed before
starting to move over other images and changing how the production
builds work.

Change-Id: I019a7c0dadafe65cb9ea9cd75bf314fb49890fb3
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/266
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Justin Miron <justin@readyset.io>