Speed up Docker build with new cargo chef
After looking at the latest version of [cargo-chef][0] I found a number
of additional [options][1] which can be passed in.

Adding the `--check` option reduces the number of compiles and splits
them between the step instead of in build. Also `--workspace` seems to
help it better understand that this is a workspace build too.

[0]: https://crates.io/crates/cargo-chef
[1]: https://github.com/LukeMathWalker/cargo-chef/blob/main/src/main.rs#L57