Fix NULL match arm bug
We have a match arm for mysql_common::Value::NULL in `to_mysql_bin` impl
for mysql_common::Value that's never supposed to be hit. This is because
we try to detect null values prior to this and write a Null bit map, as
per MySQL convention to save space w/ null values.

Unfortunately we did hit this match arm. We only call the method
`to_mysql_bin` in one location and before we do that we first check if
it's null or not by calling the `is_null` method. This should ensure
that we never hit this match arm, because we would have already written
out a null bit map.

What happened is that the default trait impl for `is_null` is to return
`false`. We didn't overwrite that trait impl for our implementation for
&T. So in that case we would return false always, and then deref and
call down to the true `to_mysql_bin` method where we discovered we
actually were null.

This fixes this by making sure that our &T impl overrides `is_null` by
derefing self and then calling down to the true `is_null` method.

This was caught by the gorm framework test when we tried to write back a
column that was NULL. For more clarity around this, gorm doesn't hard
delete data, but rather soft deletes it by adding a hidden `deleted_at`
column to every table that defaults to NULL.

Change-Id: Ib9a83c2d4c759bb514ca5d694ed4219ee20a6064
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/270
Tested-by: Buildkite CI
Reviewed-by: Michael Cronce <mc@readyset.io>
Reviewed-by: Griffin Smith <griffin@readyset.io>