Validate CI common steps against git diff config
I noticed while running a number of ops builds that a couple of
additional things were running that were not related. This validator
makes sure that the common steps are reflected and we don't miss any as
we rename things.

This also fixes the couple that were missed. This led me to finding
there were weird side effects on things and so I reorganized the
buildkite.yaml to make it clearer and separate out things.

Change-Id: I7b1edb0533608e6831c25e49a8ab16e660c5a4c9
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/269
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>